working on api.js to get data from glassdoor api. might switch to fetch from axios (issue with headers)
